learn

+++ src/parser.cpp
@@ -8,7 +8,7 @@ Parser::Parser() : document(NULL), doc_length(0), doc_position(0), line_number(1
 
 Parser::~Parser()
 {
-	for(int i=0; i<lexeme_stream.size(); i++)
+	for(size_t i=0; i<lexeme_stream.size(); i++)
 	{
 		if(lexeme_stream[i].lex_type == T_STRING)
 			delete [] lexeme_stream[i].lex_string;
@@ -22,7 +22,7 @@ void Parser::parse(const char* doc)
 	if(document != NULL)
 	{
 		json.clear();
-		for(int i=0; i<lexeme_stream.size(); i++)
+		for(size_t i=0; i<lexeme_stream.size(); i++)
 		{
 			if(lexeme_stream[i].lex_type == T_STRING)
 				delete [] lexeme_stream[i].lex_string;
